regorganized a bit
