+++ client/unit-tests/symbolic/test_modifiers.py
@@ -1,12 +1,13 @@
 import pytest
 
 from valor.symbolic.modifiers import (
+    Equatable,
+    Listable,
+    Nullable,
+    Quantifiable,
+    Spatial,
     Symbol,
     Variable,
-    Equatable,
-    Quantifiable,
-    Nullable,
-    Spatial,
 )
 
 
@@ -74,60 +75,57 @@ def test_variable():
         Variable.definite("hello")
 
 
-def test_equatable():
-    varA = Equatable.symbolic(name="A")
-    varB = Equatable.symbolic(name="B")
-    varC = Equatable.symbolic(name="C")
+def _test_equatable(varA, varB, varC):
 
-    # Equals
+    # equal
     assert (varA == varB).to_dict() == {
         "op": "eq",
         "lhs": {
             "type": "symbol",
             "value": {
                 "owner": None,
-                "name": "A",
+                "name": "a",
                 "key": None,
                 "attribute": None,
-            }
+            },
         },
         "rhs": {
             "type": "symbol",
             "value": {
                 "owner": None,
-                "name": "B",
+                "name": "b",
                 "key": None,
                 "attribute": None,
-            }
-        }
+            },
+        },
     }
     assert (varA == varB).to_dict() == (varA == Symbol("B")).to_dict()
 
-    # Not Equals
+    # not equal
     assert (varA != varB).to_dict() == {
         "op": "ne",
         "lhs": {
             "type": "symbol",
             "value": {
                 "owner": None,
-                "name": "A",
+                "name": "a",
                 "key": None,
                 "attribute": None,
-            }
+            },
         },
         "rhs": {
             "type": "symbol",
             "value": {
                 "owner": None,
-                "name": "B",
+                "name": "b",
                 "key": None,
                 "attribute": None,
-            }
-        }
+            },
+        },
     }
     assert (varA != varB).to_dict() == (varA != Symbol("B")).to_dict()
 
-    # In (exists within list)
+    # in (exists within list)
     assert varA.in_([varB, varC]).to_dict() == {
         "op": "or",
         "args": [
@@ -137,20 +135,20 @@ def test_equatable():
                     "type": "symbol",
                     "value": {
                         "owner": None,
-                        "name": "A",
+                        "name": "a",
                         "key": None,
                         "attribute": None,
-                    }
+                    },
                 },
                 "rhs": {
                     "type": "symbol",
                     "value": {
                         "owner": None,
-                        "name": "B",
+                        "name": "b",
                         "key": None,
                         "attribute": None,
-                    }
-                }
+                    },
+                },
             },
             {
                 "op": "eq",
@@ -158,36 +156,308 @@ def test_equatable():
                     "type": "symbol",
                     "value": {
                         "owner": None,
-                        "name": "A",
+                        "name": "a",
                         "key": None,
                         "attribute": None,
-                    }
+                    },
                 },
                 "rhs": {
                     "type": "symbol",
                     "value": {
                         "owner": None,
-                        "name": "C",
+                        "name": "c",
                         "key": None,
                         "attribute": None,
-                    }
-                }
-            }
-        ]
+                    },
+                },
+            },
+        ],
     }
     assert (
-        varA.in_([varB, varC]).to_dict() 
+        varA.in_([varB, varC]).to_dict()
         == varA.in_([Symbol("B"), Symbol("C")]).to_dict()
     )
 
-
-def test_quantifiable():
-    pass
-
-
-def test_nullable():
-    pass
+    # hashable
+    assert {varA, varB} == {varB, varA}
 
 
-def test_spatial():
-    pass
+def _test_quantifiable(varA, varB, varC):
+
+    # greater-than
+    assert (varA > varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "gt",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+    # greater-than or equal
+    assert (varA >= varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "ge",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+    # less-than
+    assert (varA < varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "lt",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+    # less-than or equal
+    assert (varA <= varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "le",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+
+def _test_nullable(varA, varB, varC):
+    # is none
+    assert varA.is_none().to_dict() == {
+        "arg": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "isnull",
+    }
+
+    # is not none
+    assert varA.is_not_none().to_dict() == {
+        "arg": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "isnotnull",
+    }
+
+
+def _test_spatial(varA, varB, varC):
+    # intersects
+    assert varA.intersects(varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "intersects",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+    # inside
+    assert varA.inside(varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "inside",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+    # outside
+    assert varA.outside(varB).to_dict() == {
+        "lhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "a",
+                "owner": None,
+            },
+        },
+        "op": "outside",
+        "rhs": {
+            "type": "symbol",
+            "value": {
+                "attribute": None,
+                "key": None,
+                "name": "b",
+                "owner": None,
+            },
+        },
+    }
+
+
+def _test_listable(varA, varB, varC):
+    listA = varA.list().symbolic("list_of_A")
+    assert listA.to_dict() == {
+        "type": "symbol",
+        "value": {
+            "attribute": None,
+            "key": None,
+            "name": "list_of_a",
+            "owner": None,
+        },
+    }
+
+
+def test_modifiers():
+
+    # equatable
+    A = Equatable.symbolic("A")
+    B = Equatable.symbolic("B")
+    C = Equatable.symbolic("C")
+    _test_equatable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_quantifiable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_nullable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_spatial(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_listable(A, B, C)
+
+    # quantifiable
+    A = Quantifiable.symbolic("A")
+    B = Quantifiable.symbolic("B")
+    C = Quantifiable.symbolic("C")
+    _test_equatable(A, B, C)
+    _test_quantifiable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_nullable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_spatial(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_listable(A, B, C)
+
+    # nullable
+    A = Nullable.symbolic("A")
+    B = Nullable.symbolic("B")
+    C = Nullable.symbolic("C")
+    _test_nullable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_equatable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_quantifiable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_spatial(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_listable(A, B, C)
+
+    # spatial
+    A = Spatial.symbolic("A")
+    B = Spatial.symbolic("B")
+    C = Spatial.symbolic("C")
+    _test_spatial(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_equatable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_quantifiable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_nullable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_listable(A, B, C)
+
+    # listable
+    A = Listable.symbolic("A")
+    B = Listable.symbolic("B")
+    C = Listable.symbolic("C")
+    _test_listable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_equatable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_quantifiable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_nullable(A, B, C)
+    with pytest.raises(AttributeError):
+        _test_spatial(A, B, C)
